Add validation to deliverable partnership delete

+++ marlo-web/src/main/java/org/cgiar/ccafs/marlo/data/manager/impl/DeliverablePartnershipManagerImpl.java
@@ -174,9 +174,11 @@ public class DeliverablePartnershipManagerImpl implements DeliverablePartnership
     List<DeliverablePartnership> deliverablePartnerships =
       deliverablePartnershipDAO.findByDeliverablePhasePartnerAndPartnerperson(deliverableID, phase.getId(),
         projectPartnerId, projectPartnerPersonId, partnerDivisionId, partnerType);
-    if (deliverablePartnerships != null) {
+    if (deliverablePartnerships != null && !deliverablePartnerships.isEmpty()) {
       for (DeliverablePartnership dePartnership : deliverablePartnerships) {
-        deliverablePartnershipDAO.deleteDeliverablePartnership(dePartnership.getId());
+        if (dePartnership.getId() != null) {
+          deliverablePartnershipDAO.deleteDeliverablePartnership(dePartnership.getId());
+        }
       }
     }
